Bugfix: Use replace during directory iteration
During directory iteration the content of the underlying cached content
is moerged with the working copy content. However, instead of replace,
insert was used to add the overlay content which led to changed files
being dropped from the directory that is then added to the repository.

A test is missing for that, therefore I added a comment and an
"unimplemented!" to where this should be tested.