Pin the workspace head to current commit in the presence of changes
If the workspace contains changes, we must not move away from the
current base commit. Otherwise, we might throw away changes by other
people. At the same time we want to get the latest HEAD if there are no
local changes. Therefore, we pin the underlying cache to the current
HEAD as soon as we open a file for writing and unpin it as soon as the
commit was done and no changed files are kept open.