Add support for merging changes into base branch
If something changed upstream, a ref update will fail since it is not a
fast forward ref update. In this case we try to merge the commit into
the HEAD revision of the ref. If this also fails, we'd need to merge
locally as we have a conflict. This needs to be implemented in a future
change.