Replace commit file variant in SpecialFile
This requires one major change in the way the buffered file is handled:
Instead of implicitly assuming that there is a way to get the initial
content, we enable SpecialFiles to associate operations on the
ControlDir that react on events during the file's lifecycle
(SpecialFileOps). They now exist for the three currently existing
SpecialFiles.