Lock and update open working copy files after merges
If there is an open file in the working copy, we need to lock these
files so that there are no writes during the commit, thus avoiding race
conditions. After that, we need to update the contents of these files
since there could have been merges that alter these files.