Consolidate impls of Directory and add find function
There were several impls based on Vec<DirectoryEntry>. They are now
all based directly on Vec<DirectoryEntry> where possible. In addition,
there is now a function to search for an entry. This way, it is possible
to provide a hash or tree based directory to improve the speed for
larger directories.