Bugfix: Return correct attribute timeout in getattr
During testing an inconsistency was observed after reverting a changed file:
Whenever the file was shrunk by deleting content and these changes were
reverted, the file was truncated by the amount of bytes that had been removed
previously removed from the file. On the second open, the file was back to its
original state.

This behavior was caused by the kernel not querying the new file size after the
revert. The data that is then read after the first open call after the revert is
then truncated to the size as it was before the revert, causing missing data at
the end of the file.

The reason for the missing getattr call is a cache inside the Linux kernel that
caches the file's attribute. The caching duration is determined by the result of
the getattr call: On success, getattr returns a tuple that contains a timespec.
The assumption was that this was the time when the attributes were generated.
However it turned out that its a relative time tht reflects the ttl of the
attributes. A new constant was introduced that statically configures the timeout
in ms. This value is then returned to the kernel.